% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlFredData.R
\name{dlFredData}
\alias{dlFredData}
\title{dlFredData}
\usage{
dlFredData(sites, useChrome = TRUE, parallel = FALSE, meta = FALSE,
  output = NULL)
}
\arguments{
\item{sites}{the data frame returned by the getFred function}

\item{useChrome}{indicates if Selenium should drive an instance of a Chrome browser}

\item{parallel}{Indicator of whether or not to use parallelized function calls
when}

\item{meta}{Indicates whether or not the file meta data should be scraped and
put into an file}

\item{output}{The name of the file to save the metadata in}
}
\description{
Uses Selenium to simulate website interactions to download data
}
\examples{
\donttest{

# Set up parallel processing if wanted
useParallel <- rideR::rideParallel(8)

# Search over first 100,000 pages with IDs [0, 100000]
urls <- rideR::getFred(isParallel = useParallel)

# Download the files and save/scrape the metadata
dlFredData(urls, parallel = useParallel, meta = TRUE, output = "~/Desktop/fredMetaData.csv")
}

}

