% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFred.R
\name{getFred}
\alias{getFred}
\title{getFred}
\usage{
getFred(start = 0, end = 1e+05, candidatesOnly = TRUE, outFile = NULL,
  outObject = NULL, isParallel = FALSE)
}
\arguments{
\item{start}{Defines the FileID that the function will begin the search}

\item{end}{Defines the FileID that the function will end the search}

\item{candidatesOnly}{a Boolean used to return only the URLs that contain
a link used to trigger options for downloading the associated data}

\item{outFile}{A file path where a csv containing the data will be written to
(useful if the data will be shared with others and/or to maintain a list in
MS Excel or a database).}

\item{outObject}{A file path where an R object will be stored (most useful if
you need to run parts of the process at different times)}

\item{isParallel}{a Boolean passed to internal function calls used to process
the data in parallel}
}
\value{
A data frame containing the URLs and Booleans indicating if the URL
contains the link used to trigger download options
}
\description{
A brute force approach to identifying candidate URLs that may
have Frequently Requested Educational Data files for download.
}
\examples{
\donttest{

# Check the range of sites that may have valid data download links from the
# Frequently Requested Educational Data web portal.  This will check for
# all sites with fileID values of 0 - 100,000
sites <- rideR::getFred()

}
}

