% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCardCleaner.R
\name{reportCardCleaner}
\alias{reportCardCleaner}
\title{reportCardCleaner}
\usage{
reportCardCleaner(page, school = TRUE)
}
\arguments{
\item{page}{the URL whose content is to be parsed}

\item{school}{Whether the pages being parsed are school report cards}
}
\value{
A dataframe containing the report card website contents
}
\description{
Used to scrape and organize contents of report card websites
}
\examples{
\donttest{
# Get report card content quickly
para <- rideR::rideParallel(8)

# Gets the URLs with the report card data
rideAMO <- rideReportCards(para)

# Get all of the report card data for Districts
distData <- plyr::ldply(rideAMO[[1]], .fun = reportCardCleaner, school = FALSE,
						.parallel = para)

# And saves the data frame as a file that can be used elsewhere
write.csv(distData, "~/Desktop/districtReportCards.csv", row.names = FALSE)

# Get all of the report card data for Districts
schData <- plyr::ldply(rideAMO[[2]], .fun = reportCardCleaner, .parallel = para)

# And saves the data frame as a file that can be used elsewhere
write.csv(schData, "~/Desktop/schoolReportCards.csv", row.names = FALSE)

}
}

