% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rideParallel.R
\name{rideParallel}
\alias{rideParallel}
\title{Setup Parallel}
\usage{
rideParallel(threads = NULL)
}
\arguments{
\item{threads}{The number of "cores" (or threads) to use when parallelizing processes}
}
\value{
An indicator that can be passed to the .parallel argument of dplyr methods
}
\description{
Sets up parallel processing.  Used if called from a computer running OSX
}
\examples{
\donttest{
# Sets the number of cores to use for parallelized function calls
# This only has an effect on non-MS operating systems where the doMC package
# can be installed
para <- rideR::rideParallel(32)

}
}

