% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rimaParallel.R
\name{rimaParallel}
\alias{rimaParallel}
\title{Setup Parallel}
\usage{
rimaParallel(threads = NULL)
}
\arguments{
\item{threads}{The number of "cores" (or threads) to use when parallelizing processes}
}
\value{
An indicator that can be passed to the .parallel argument of dplyr methods
}
\description{
Sets up parallel processing.  Used if called from a computer running OSX
}

