% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rideReportCards.R
\name{rideReportCards}
\alias{rideReportCards}
\title{rideReportCards}
\usage{
rideReportCards(parallel = FALSE, distout = NULL, schout = NULL)
}
\arguments{
\item{parallel}{Indicator of whether or not to use parallelized function calls
when}

\item{distout}{The name of the file where the district level html will be saved}

\item{schout}{The name of the file where the school level html will be saved}
}
\value{
A list containing the list of district and list of school HTML content
}
\description{
Scrapes ESEA accountability report card data from RIDE
}
\examples{
\donttest{
# Get report card content quickly
rideAMO <- rima::rimaParallel(8) \%>\% rideReportCards()
}
}

